Created backup system, Added more blacklisted dimensions

+++ src/main/java/net/willsr71/mystcraftutils/WorldManager.java
@@ -1,9 +1,11 @@
 package net.willsr71.mystcraftutils;
 
+import org.apache.commons.io.FileUtils;
 import org.bukkit.Bukkit;
 import org.bukkit.World;
 
 import java.io.File;
+import java.io.IOException;
 
 public class WorldManager {
     private MystcraftUtils plugin;
@@ -15,7 +17,11 @@ public class WorldManager {
     public void deleteWorld(String dimension){
         World world = Bukkit.getWorld(dimension);
         File worldFolder = world.getWorldFolder();
-        Bukkit.unloadWorld(dimension, true);
+        boolean success = Bukkit.unloadWorld(dimension, true);
+        if(!copyFolder(worldFolder, dimension)){
+            plugin.getLogger().warning("Error copying dimension folder " + worldFolder.getAbsolutePath());
+            return;
+        }
         deleteFolder(worldFolder);
     }
 
@@ -31,5 +37,17 @@ public class WorldManager {
             }
         }
         return(path.delete());
+    }
+
+    private boolean copyFolder(File path, String name){
+        File backupFolder = new File("worldBackups/" + name);
+        backupFolder.mkdirs();
+        try{
+            FileUtils.copyDirectory(path, backupFolder);
+            return true;
+        }catch (IOException e){
+            e.printStackTrace();
+        }
+        return false;
     }
 }
